This is another improvement to `serde_v8`'s serialization code, it drops the need for creating a `v8::EscapableHandleScope` or "subscope". It's a simplication and optimization similar in nature to #9987.

This is very exciting since sub 100ns/op for sync-ops appears to be within reach ! (at least on my machine, M1 Air)

## Benches

Using our trusty `op_baseline` bench to compare with `main`:
(Note that the before and after of these benches integrates changes from #9983)

```
Before:
test bench_op_async   ... bench:     647,272 ns/iter (+/- 57,528)
test bench_op_nop     ... bench:     157,173 ns/iter (+/- 2,096)
test bench_op_pi_bin  ... bench:     179,239 ns/iter (+/- 4,354)
test bench_op_pi_json ... bench:     175,075 ns/iter (+/- 1,814)

After:
test bench_op_async   ... bench:     564,882 ns/iter (+/- 12,229)
test bench_op_nop     ... bench:     114,117 ns/iter (+/- 2,489)
test bench_op_pi_bin  ... bench:     139,160 ns/iter (+/- 6,047)
test bench_op_pi_json ... bench:     129,266 ns/iter (+/- 2,938)
```